+++ rvp64_scan/builtin-rvp64-scmplt8.c
@@ -9,7 +9,7 @@
 #include <stdint.h>
 
 static __attribute__ ((noinline))
-uint32_t tlpmcs (uint32_t ra, uint32_t rb)
+uint64_t tlpmcs (uint64_t ra, uint64_t rb)
 {
   return __rv__scmplt8 (ra, rb);
 }
